requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# OAuth and security
httpx==0.26.0
oauthlib==3.2.2
requests==2.31.0
cryptography==42.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# AWS S3
boto3==1.34.34

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
